Render applied template the same way, with time
Ensures the HTML sent over the wire when a job application is
created via AJAX is the same as the one that appears in the
Postings listing (index) when that page is loaded from the server.

Also includes a migration to convert date_sent to DateTime instead
of just a Date in the future, allowing more accurate application
times to be used ("applied five minutes ago" over "applied today").