Fix deploys; enable deployment on new chef configs
This replaces the existing deploy configs with new ones matched
to the chef configs I created. It fixes a lot of issues and allows
us to run but a few commands to create the server and deploy to it.

This also allows us to run multiple apps per-server, which could
even include a staging and production instance on the same server,
though we won't do that for now. ^_^

* Partially addresses #79 by automating a LARGE amount of the
  deployment process, but I won't tick this one off quite yet as
  there are still like two manual things that need be done on
  first run
* Fix 90 (this allows us to easily set up a staging site with a
  separate capistrano config in `config/deploy/staging.rb`)
* Fix 140 (this runs migrations on every deploy)
* Fix 141 (this runs bundle install on every deploy)